Back in a testable state!
Let's verify everything that we DO have works, then we can finish the rest of the opcodes.

It turns out that INCLUDES need to be at the bottom of the program, right before the END command. Then any INCLUDE files with their own INCLUDES need to put them at the bottom.